split tests: test_pipes.py
